:sparkles: add selfreboot job in d05-acpi-weekly.json and add config in 138.conf
In d05-acpi-weekly-template.json:

1.add pre_autotest and end_autotest job for contructing test environment
befor test and recover test environment after test

2.add boot_linaro_image job before running any plinth test to reboot
the board for update the test Image

In hip-7-d05-138.conf

1.add boot_cmds_disc_selfreboot descript in 138.conf.but 138 is not
support selfreboot and xge,roce,perf is not excute in 138.so this
descript is the same as boot_cmds_disc_acpi which run a pxe reboot but not
a disc reboot.

:sparkles: modify d05 config files to adapt to new bios

in hip07-d05-159.conf

1.add sendcontrol "[" before expect "Move Highlight" in disc_acpi
because bios need user to input Esc to enter bios UI except any button

2.change the selfreboot mothod from boot manager to boot from file

in hip05-d05-138.conf

1.add sendcontrol "[" before expect "Move Highlight" in disc_acpi
because bios need user to input Esc to enter bios UI except any button

in hip07-d05
1.modify interrupt_boot_promt from seconds to stop automatical booting
to F2 to start setup for bios changge the description.